Updates example after e2e test
